Refactor last unnecessary global components
