+++ src/components/cards/TradeCard/TradePair.vue
@@ -146,7 +146,7 @@ import useNumbers from '@/composables/useNumbers';
 import { ETHER } from '@/constants/tokenlists';
 
 import TradePairToggle from '@/components/cards/TradeCard/TradePairToggle.vue';
-import SelectTokenModal from '@/components/modals/SelectTokenModal.vue';
+import SelectTokenModal from '@/components/modals/SelectTokenModal/SelectTokenModal.vue';
 
 const ETH_BUFFER = 0.1;
