+++ src/components/modals/SelectTokenModal/Search.vue
