+++ src/components/modals/SelectTokenModal/SelectTokenModal.vue
@@ -25,7 +25,7 @@
           v-if="Object.keys(tokenLists).length > 0"
           class="h-96 overflow-y-scroll"
         >
-          <RowTokenlist
+          <TokenListsListItem
             v-for="(tokenList, i) in tokenLists"
             :key="i"
             :isActive="isActiveList(tokenList.name)"
@@ -71,7 +71,7 @@
           :buffer="100"
         >
           <a @click="onSelectToken(item.address)">
-            <RowToken :token="item" />
+            <TokenListItem :token="item" />
           </a>
         </RecycleScroller>
         <div
@@ -99,8 +99,17 @@ import { useI18n } from 'vue-i18n';
 import { clone } from '@/utils';
 import { isAddress, getAddress } from '@ethersproject/address';
 import useTokenLists from '@/composables/useTokenLists';
+import TokenListItem from '@/components/lists/TokenListItem.vue';
+import TokenListsListItem from '@/components/lists/TokenListsListItem.vue';
+import Search from './Search.vue';
 
 export default defineComponent({
+  components: {
+    TokenListItem,
+    TokenListsListItem,
+    Search
+  },
+
   emits: ['close', 'selectTokenlist', 'select'],
 
   props: {
