+++ src/components/pages/pool/PoolActivitiesCard/Activities.vue
@@ -1,5 +1,5 @@
 <template>
-  <TablePoolActivities
+  <Table
     :tokens="pool ? pool.tokensList : []"
     :pool-activities="poolActivities"
     :is-loading="loading || isLoadingPoolActivities"
@@ -19,6 +19,8 @@ import { defineComponent, computed, PropType } from 'vue';
 import { useRoute } from 'vue-router';
 import { flatten } from 'lodash';
 
+import Table from './Table.vue';
+
 import usePoolActivitiesQuery from '@/composables/queries/usePoolActivitiesQuery';
 import usePoolUserActivitiesQuery from '@/composables/queries/usePoolUserActivitiesQuery';
 
@@ -27,6 +29,10 @@ import { FullPool } from '@/services/balancer/subgraph/types';
 import { PoolActivityTab } from './types';
 
 export default defineComponent({
+  components: {
+    Table
+  },
+
   props: {
     pool: {
       type: Object as PropType<FullPool>,
